+++ src/parser/api.ts
@@ -1,4 +1,3 @@
 ﻿export { parse } from './api/parse';
 export { bind } from './api/bind';
-export { escape } from './api/escape';
 export { caches } from './api/cache';