+++ src/renderer/render/media/slideshare.ts
@@ -1,4 +1,4 @@
-﻿import { parse, escape } from '../../../parser';
+﻿import { parse } from '../../../parser';
 import { cache } from '../../../parser/inline/media';
 import { sanitize } from 'dompurify';
 import DOM, { html, define } from 'typed-dom';
@@ -26,7 +26,7 @@ export function slideshare(url: URL): HTMLElement | undefined {
       },
       error({ status, statusText }) {
         assert(Number.isSafeInteger(status));
-        void define(outer, parse(`*[]( ${url.href} )\\\n-> ${status}: ${escape(statusText)}*`).querySelector('p')!.childNodes);
+        void define(outer, [parse(`*[]( ${url.href} )*\n\n\`\`\`\n${status}\n${statusText}\n\`\`\``)]);
       },
     });
     return outer;